requirements.txt
flask==3.0.2
pillow==10.2.0
requests==2.31.0
